+++ lib/src/functions/math.dart
@@ -289,11 +289,28 @@ final _random = math.Random();
 
 final _randomFunction = _function("random", r"$limit: null", (arguments) {
   if (arguments[0] == sassNull) return SassNumber(_random.nextDouble());
-  var limit = arguments[0].assertNumber("limit").assertInt("limit");
-  if (limit < 1) {
+  var limit = arguments[0].assertNumber("limit");
+
+  if (limit.hasUnits) {
+    warn(
+      "math.random() will no longer ignore \$limit units ($limit) in a "
+      "future release.\n"
+      "\n"
+      "Recommendation: "
+      "math.random(math.div(\$limit, 1${limit.unitString})) * 1${limit.unitString}\n"
+      "\n"
+      "To preserve current behavior: "
+      "math.random(math.div(\$limit, 1${limit.unitString}))\n"
+      "\n"
+      "More info: https://sass-lang.com/d/random-with-units",
+    );
+  }
+
+  var limitScalar = limit.assertInt("limit");
+  if (limitScalar < 1) {
     throw SassScriptException("\$limit: Must be greater than 0, was $limit.");
   }
-  return SassNumber(_random.nextInt(limit) + 1);
+  return SassNumber(_random.nextInt(limitScalar) + 1);
 });
 
 final _div = _function("div", r"$number1, $number2", (arguments) {